feat: move handler into separate folder
