+++ src/handler/mod.rs
@@ -1,3 +1,4 @@
+#[warn(missing_debug_implementations, rust_2018_idioms, missing_docs)]
 use crate::context::Context;
 use crate::error::ObsidianError;
