+++ src/lib.rs
@@ -4,10 +4,10 @@ mod app;
 pub mod error;
 
 pub mod context;
+pub mod handler;
 pub mod middleware;
 pub mod router;
 
 pub use app::{App, EndpointExecutor};
 pub use error::ObsidianError;
 pub use hyper::{header, Body, HeaderMap, Method, Request, Response, StatusCode, Uri, Version};
-pub use router::ContextResult;
